move migration assistant viewmodel to its namesake namespace
