+++ src/PythonMigrationViewExtension/Controls/VisualDifferenceViewer.xaml
@@ -1,12 +1,12 @@
-﻿<Window x:Class="Dynamo.PythonMigration.Controls.VisualDifferenceViewer"
+﻿<Window x:Class="Dynamo.PythonMigration.Dynamo.PythonMigration.MigrationAssistant.Controls.VisualDifferenceViewer"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:p="clr-namespace:Dynamo.PythonMigration.Properties;assembly=PythonMigrationViewExtension"
+        xmlns:p="clr-namespace:Dynamo.PythonMigration.Dynamo.PythonMigration.MigrationAssistant.Properties;assembly=PythonMigrationViewExtension"
         xmlns:fa="http://schemas.fontawesome.io/icons/"
         xmlns:ui="clr-namespace:Dynamo.UI;assembly=DynamoCoreWpf"
-        xmlns:local="clr-namespace:Dynamo.PythonMigration.Controls"
+        xmlns:local="clr-namespace:Dynamo.PythonMigration.Dynamo.PythonMigration.MigrationAssistant.Controls"
         xmlns:diffplex="clr-namespace:DiffPlex.Wpf.Controls;assembly=DiffPlex.Wpf"
         mc:Ignorable="d"
         d:DesignHeight="450"
