+++ src/PythonMigrationViewExtension/IronPythonInfoDialog.xaml
@@ -1,8 +1,8 @@
-﻿<Window x:Class="Dynamo.PythonMigration.IronPythonInfoDialog"
+﻿<Window x:Class="Dynamo.PythonMigration.Dynamo.PythonMigration.MigrationAssistant.IronPythonInfoDialog"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:ui="clr-namespace:Dynamo.UI;assembly=DynamoCoreWpf"
-        xmlns:w="clr-namespace:Dynamo.PythonMigration.Properties"
+        xmlns:w="clr-namespace:Dynamo.PythonMigration.Dynamo.PythonMigration.MigrationAssistant.Properties"
         Width="660"
         SizeToContent="Height"
         ResizeMode="NoResize"
